+++ Changelog.md
@@ -10,7 +10,12 @@
 - Added Laravel 12 Support
 - BC: Added the `getBuyable()` method to the `OrderItem` interface
 - BC: Added the `addSubItem()` method to the Cart interface
-- BC: Added the `hasParent()` and `getParent()` methods to the CartItem interface
+- BC: Added the `$forceNewItem` (default false) parameter to the `Cart::addItem()` method
+- BC: Added the following methods to the CartItem interface:
+    - `hasParent()`
+    - `getParent()`
+    - `hasChildItems()`
+    - `getChildItems()`
 - BC: Added the `getOriginalPrice()` and `hasAHigherOriginalPrice()` methods to the Buyable interface
 - BC: Added the following methods to the `ProductState` interface:
     - `isListable()`